refactor : 하위 패키지 중 domain을 model로 수정
